+++ src/main/java/urfu/picnic/entity/TeamMember.java
@@ -1,12 +1,11 @@
 package urfu.picnic.entity;
 
-import javax.persistence.*;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,16 +19,16 @@ public class TeamMember {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "memberId", nullable = false, unique = true)
-    private int memberId;
+    private Long memberId;
 
     @Column(name = "userId", nullable = false)
-    private int userId;
+    private Long userId;
 
     @Column(name = "teamId", nullable = false)
-    private int teamId;
+    private Long teamId;
 
     @Column(name = "role")
-    private int role; //роль учатсника в команде переименовать таблицу на схеме
+    private Long role; //роль учатсника в команде переименовать таблицу на схеме
 
     @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinColumn(name = "teamid")
